Move ListItem base component out of barrel
