+++ packages/app/ui/components/listItems/InteractableChatListItem.tsx
@@ -22,7 +22,7 @@ import { ColorTokens, Stack, View, getTokenValue, isWeb } from 'tamagui';
 
 import * as utils from '../../utils';
 import { ChatListItem } from './ChatListItem';
-import { ListItemProps } from './ListItem';
+import { ListItemProps } from '../ListItem';
 import { useBoundHandler } from './listItemUtils';
 
 function BaseInteractableChatRow({
