+++ packages/app/ui/components/listItems/index.tsx
@@ -3,5 +3,5 @@ export * from './ChatListItem';
 export * from './ContactListItem';
 export * from './ChannelListItem';
 export * from './InteractableChatListItem';
-export * from './ListItem';
+export * from '../ListItem';
 export * from './SystemContactListItem';
